Style : 타입 경로 변경
